ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
onnxruntime==1.16.3
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
diffusers==0.24.0
clip-by-openai==1.0
scikit-learn==1.3.2
matplotlib==3.8.2
redis==5.0.1
psycopg2-binary==2.9.9
